+++ apps/explorer/src/hooks/useAppData.ts
@@ -1,41 +1,53 @@
-import { useEffect, useMemo, useState } from 'react'
-
 import { AnyAppDataDocVersion } from '@cowprotocol/app-data'
 
-import { DEFAULT_IPFS_READ_URI } from 'const'
+import { DEFAULT_IPFS_READ_URI, IPFS_INVALID_APP_IDS } from 'const'
 import { metadataApiSDK } from 'cowSdk'
-import { useNetworkId } from 'state/network'
+import useSWR from 'swr'
 
-export const useAppData = (
-  appDataHash: string,
-  isLegacyAppDataHex: boolean
-): { isLoading: boolean; appDataDoc: AnyAppDataDocVersion | void | undefined } => {
-  const network = useNetworkId() || undefined
-  const [isLoading, setLoading] = useState<boolean>(false)
-  const [appDataDoc, setAppDataDoc] = useState<AnyAppDataDocVersion | void>()
-  useEffect(() => {
-    async function getAppDataDoc(): Promise<void> {
-      setLoading(true)
-      try {
-        const decodedAppData = await fetchDocFromAppDataHex(appDataHash, isLegacyAppDataHex)
-        setAppDataDoc(decodedAppData)
-      } catch (e) {
-        const msg = `Failed to fetch appData document`
-        console.error(msg, e)
-      } finally {
-        setLoading(false)
-        setAppDataDoc(undefined)
-      }
-    }
-    getAppDataDoc()
-  }, [appDataHash, network, isLegacyAppDataHex])
+import { decodeFullAppData } from '../utils/decodeFullAppData'
 
-  return useMemo(() => ({ isLoading, appDataDoc }), [isLoading, appDataDoc])
+interface AppDataDecodingResult {
+  isLoading: boolean
+  appDataDoc: AnyAppDataDocVersion | undefined
+  hasError: boolean
+  ipfsUri: string | undefined
+}
+
+export const useAppData = (appData: string, fullAppData?: string): AppDataDecodingResult => {
+  // Old AppData use a different way to derive the CID (we know is old if fullAppData is not available)
+  const isLegacyAppDataHex = fullAppData === undefined
+
+  const {
+    error: ipfsError,
+    isLoading: isIpfsLoading,
+    data: ipfsUri,
+  } = useSWR(['appDataHexToCid', appData, isLegacyAppDataHex], async ([_, appData, isLegacyAppDataHex]) => {
+    const cid = await appDataHexToCid(appData.toString(), isLegacyAppDataHex)
+
+    return `${DEFAULT_IPFS_READ_URI}/${cid}`
+  })
+
+  const {
+    error,
+    isLoading,
+    data: appDataDoc,
+  } = useSWR(
+    ['getDecodedAppData', appData, fullAppData, isLegacyAppDataHex],
+    async ([_, appData, fullAppData, isLegacyAppDataHex]) => {
+      const { error, decodedAppData } = await getDecodedAppData(appData, isLegacyAppDataHex, fullAppData)
+
+      if (error) throw error
+
+      return decodedAppData || undefined
+    },
+  )
+
+  return { isLoading: isLoading || isIpfsLoading, hasError: !!(ipfsError || error), appDataDoc, ipfsUri }
 }
 
 export const fetchDocFromAppDataHex = (
   appDataHex: string,
-  isLegacyAppDataHex: boolean
+  isLegacyAppDataHex: boolean,
 ): Promise<void | AnyAppDataDocVersion> => {
   const method = isLegacyAppDataHex ? 'fetchDocFromAppDataHexLegacy' : 'fetchDocFromAppDataHex'
   return metadataApiSDK[method](appDataHex, DEFAULT_IPFS_READ_URI)
@@ -44,4 +56,28 @@ export const fetchDocFromAppDataHex = (
 export const appDataHexToCid = (appDataHash: string, isLegacyAppDataHex: boolean): Promise<string | void> => {
   const method = isLegacyAppDataHex ? 'appDataHexToCidLegacy' : 'appDataHexToCid'
   return metadataApiSDK[method](appDataHash)
+}
+
+async function getDecodedAppData(
+  appData: string,
+  isLegacyAppDataHex: boolean,
+  fullAppData?: string,
+): Promise<{ decodedAppData?: void | AnyAppDataDocVersion; error?: Error }> {
+  // If the full appData is available, we try to parse it as JSON
+  if (fullAppData) {
+    try {
+      const decodedAppData = decodeFullAppData(fullAppData, true)
+      return { decodedAppData }
+    } catch (error) {
+      console.error('Error parsing fullAppData from the API', { fullAppData }, error)
+      return { error }
+    }
+  }
+
+  if (IPFS_INVALID_APP_IDS.includes(appData.toString())) {
+    return { error: new Error('Invalid app id') }
+  }
+
+  const decodedAppData = await fetchDocFromAppDataHex(appData.toString(), isLegacyAppDataHex)
+  return { decodedAppData }
 }