+++ libs/hook-dapp-lib/src/types.ts
@@ -1,13 +1,22 @@
-import type { SupportedChainId } from '@cowprotocol/cow-sdk'
+import type { ReactNode } from 'react'
+
+import { HookDappType, HookDappWalletCompatibility } from './consts'
 
 export interface CowHook {
   target: string
   callData: string
   gasLimit: string
+  dappId: string
+}
+
+export interface HookDappConditions {
+  position?: 'post' | 'pre'
+  walletCompatibility?: HookDappWalletCompatibility[]
+  supportedNetworks?: number[]
 }
 
 export interface CowHookCreation {
-  hook: CowHook
+  hook: Omit<CowHook, 'dappId'>
   recipientOverride?: string
 }
 
@@ -16,12 +25,17 @@ export interface TokenData {
 }
 
 export interface CowHookDetails extends CowHookCreation {
+  hook: CowHook
+  uuid: string
+}
+
+export interface CowHookToEdit extends CowHookCreation {
   uuid: string
 }
 
 export interface CoWHookDappActions {
   addHook(payload: CowHookCreation): void
-  editHook(payload: CowHookDetails): void
+  editHook(payload: CowHookToEdit): void
   setSellToken(token: TokenData): void
   setBuyToken(token: TokenData): void
 }
@@ -33,11 +47,23 @@ export interface HookDappOrderParams {
 }
 
 export interface HookDappContext {
-  chainId: SupportedChainId
+  chainId: number
   account?: string
   orderParams: HookDappOrderParams | null
   hookToEdit?: CowHookDetails
   isSmartContract: boolean | undefined
   isPreHook: boolean
   isDarkMode: boolean
+}
+
+export interface HookDappBase {
+  id: string
+  name: string
+  descriptionShort?: string
+  description?: ReactNode | string
+  type: HookDappType
+  version: string
+  website: string
+  image: string
+  conditions?: HookDappConditions
 }